feat : features/myprofile 1차 작업 완료
