+++ client/pages/myProfile/index.tsx
@@ -1,19 +1,1 @@
-import styled from 'styled-components';
-
-import CheckMyInFo from '@/components/MyProfile/CheckMyInFo';
-
-import { styleMixin } from '@/src/shared/styles';
-
-const myProfile = () => (
-  <MyPageStyle>
-    <CheckMyInFo />
-  </MyPageStyle>
-);
-
-export default myProfile;
-
-const MyPageStyle = styled.div`
-  ${styleMixin.Column()}
-  width: 100%;
-  height: 100%;
-`;
+export { default } from '@/src/pages/myprofile';
