+++ client/src/features/myprofile/hooks/index.ts
@@ -1,0 +1,2 @@
+export { default as useMyProfileLogic } from './useMyProfileLogic';
+export { default as useMyProfileManager } from './useMyProfileManager';
