+++ client/src/features/myprofile/index.ts
@@ -1,0 +1,1 @@
+export { default as CheckMyInFo } from './CheckMyInFo';
